requirements.txt
Django==1.11.1
mysqlclient==1.3.10
django-autocomplete-light==3.2.10
django-bootstrap4==0.0.4
django-registration==2.2
django-braces==1.11.0
django-gravatar2==1.4.2
rules==1.2.1
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
coreapi==2.3.1
pygments==2.2.0
channels==2.1.1
uwsgi==2.0.15
pyyaml==3.12
